python/chatbot-memory/requirements.txt
hyrex
python-dotenv
psycopg[binary]
pgvector
fastapi
uvicorn[standard]
openai
numpy
